ISSUE #2756 split revision routes into it's own file
